Introduce a --forcemono option in the server
The purpose of the --forcemono option is to make the server process
its input only in mono even if the source sends in stereo.  While this
still allows for Mono-in/Stereo-out processing, it records the tracks
in mono (instead of stereo) and will reduce the processing
requirements a bit.  This is more or less a stopgap measure: a real
fix would have the clients signal Mono-in/stereo-out settings to the
server and send only mono streams, and also have the server signal
forced mono processing to the clients to keep them from sending
stereo (with the associated performance and bandwidth costs) when it
is not going to get processed anyway.